# Fix operator precedence in the parser