@sjakobi 2019-10-23
This does seem to impact parser performance though: The `Issue #108/Text` benchmark now takes ~7.8ms instead of ~6.9ms. Not sure whether that's acceptable…

We could consider putting this code behind a feature flag and mostly use it to discover more interesting formatting idempotence issues.