@sjakobi 2019-10-23
@Gabriel439 Your comment https://github.com/dhall-lang/dhall-haskell/pull/1460#issuecomment-545256369 seems to apply this PR:

> Yeah, the reason it's grouped like that is a performance optimization since you only run into expressions with the wrong precedence if they are ill-typed expressions

What do you think about the idea of putting the this code behind a cabal flag? The idempotence test is fairly useless as long as it can stumble over these operator precedence issues.