@Gabriel439 2019-10-23
@sjakobi: The benchmark I'm most interested in is this one:

https://github.com/dhall-lang/dhall-haskell/blob/16e6a12c3da1966b095cc8089d8c1274c8fca4e5/dhall/benchmark/parser/Main.hs#L98

That's the one where I'd expect this optimization to have the largest impact.  What is the difference that you observe for that benchmark?