@sjakobi 2019-10-23
Previously 346 μs, now ~442 μs. I guess that's a bit much! :/