@sjakobi 2019-10-23
I've tried to measure the impact on real code using `cpkg`'s `pkgs-set.dhall`:

#### `master`

```
$ bench "dhall resolve --no-cache --file pkgs/pkg-set.dhall  --immediate-dependencies > /dev/null"
benchmarking dhall resolve --no-cache --file pkgs/pkg-set.dhall  --immediate-dependencies > /dev/null
time                 173.0 ms   (172.1 ms .. 173.7 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 172.6 ms   (172.1 ms .. 172.9 ms)
std dev              498.9 μs   (143.1 μs .. 734.1 μs)
variance introduced by outliers: 12% (moderately inflated)
```

#### This branch

```
$ bench "dhall resolve --no-cache --file pkgs/pkg-set.dhall  --immediate-dependencies > /dev/null"
benchmarking dhall resolve --no-cache --file pkgs/pkg-set.dhall  --immediate-dependencies > /dev/null
time                 187.6 ms   (185.5 ms .. 189.0 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 188.5 ms   (187.9 ms .. 189.3 ms)
std dev              955.5 μs   (721.7 μs .. 1.153 ms)
variance introduced by outliers: 14% (moderately inflated)
```

Is there a better command for measuring parser performance?