@Gabriel439 2019-10-23
@sjakobi: Actually, I think that's fine.  Sure it's a measurable performance regression but it's not that bad and now it's more correct.  After all, the contract of `dhall format` is that it should work even for ill-typed code, so if it misformats such code then it's not working as advertised.  I just wanted to make sure that we measured the performance regression; that's all.